+++ packages/zowe-explorer/__tests__/__unit__/trees/job/JobFSProvider.unit.test.ts
@@ -17,6 +17,7 @@ import { ZoweExplorerApiRegister } from "../../../../src/extending/ZoweExplorerA
 import { JobFSProvider } from "../../../../src/trees/job/JobFSProvider";
 import { MockedProperty } from "../../../__mocks__/mockUtils";
 import { AuthUtils } from "../../../../src/utils/AuthUtils";
+import { Profiles } from "../../../../src/configuration/Profiles";
 
 const testProfile = createIProfile();
 
@@ -206,6 +207,13 @@ describe("createDirectory", () => {
 });
 
 describe("fetchSpoolAtUri", () => {
+    const loadNamedProfileMock = jest.fn().mockReturnValue(testProfile);
+    beforeEach(() => {
+        jest.spyOn(Profiles, "getInstance").mockReturnValue({ loadNamedProfile: loadNamedProfileMock } as any);
+    });
+    afterAll(() => {
+        jest.restoreAllMocks();
+    });
     it("fetches the spool contents for a given URI - downloadSingleSpool", async () => {
         const lookupAsFileMock = jest
             .spyOn(JobFSProvider.instance as any, "_lookupAsFile")
@@ -246,6 +254,27 @@ describe("fetchSpoolAtUri", () => {
         const lookupAsFileMock = jest
             .spyOn(JobFSProvider.instance as any, "_lookupAsFile")
             .mockReturnValueOnce({ ...testEntries.spool, data: new Uint8Array(), encoding: { kind: "other", codepage: "IBM-1147" } });
+        const newData = "spool contents";
+        const mockJesApi = {
+            downloadSingleSpool: jest.fn((opts) => {
+                opts.stream.write(newData);
+            }),
+        };
+        const jesApiMock = jest.spyOn(ZoweExplorerApiRegister, "getJesApi").mockReturnValueOnce(mockJesApi as any);
+        const entry = await JobFSProvider.instance.fetchSpoolAtUri(testUris.spool);
+        expect(mockJesApi.downloadSingleSpool).toHaveBeenCalledWith(
+            expect.objectContaining({ jobFile: testEntries.spool.spool, encoding: "IBM-1147", binary: false })
+        );
+        expect(entry.data.toString()).toStrictEqual(newData.toString());
+        jesApiMock.mockRestore();
+        lookupAsFileMock.mockRestore();
+    });
+
+    it("fetches the spool contents for a given URI - downloadSingleSpool w/ profile encoding", async () => {
+        const lookupAsFileMock = jest
+            .spyOn(JobFSProvider.instance as any, "_lookupAsFile")
+            .mockReturnValueOnce({ ...testEntries.spool, data: new Uint8Array() });
+        loadNamedProfileMock.mockReturnValueOnce({ ...testProfile, profile: { ...testProfile.profile, encoding: "IBM-1147" } });
         const newData = "spool contents";
         const mockJesApi = {
             downloadSingleSpool: jest.fn((opts) => {